Move uploadedDocs array to the component
